requirements.txt
flask==3.0.3
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
anthropic==0.44.0
flask-cors==5.0.1
